+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.19.0] - Not released
+#### Changed
+- Site admin's email added to the frozen user's error message.
 
 ## [2.18.2] - 2024-02-18
 ### Fixed